requirements.txt
matplotlib==3.5.1
pandas==1.3.5
Pillow==8.4.0
plotly==5.4.0
streamlit==1.2.0
